refactor(nx-governance): Refine Extension Host boundary by moving Nx-specific discovery/config loading out of Core-facing extension modules
