+++ packages/governance/src/nx-host/extensions/config.ts
